use eoapi structure to accommodate more complex runtime
