+++ cdk/dockerfiles/Dockerfile.raster
@@ -1,0 +1,27 @@
+ARG PYTHON_VERSION=3.11
+
+FROM public.ecr.aws/lambda/python:${PYTHON_VERSION}
+
+# Install system dependencies to compile (numexpr)
+RUN yum install -y gcc-c++
+
+WORKDIR /tmp
+RUN python -m pip install pip -U
+
+COPY runtimes/eoapi/raster /tmp/raster
+RUN python -m pip install "mangum>=0.14,<0.15" /tmp/raster["psycopg-binary"] -t /asset --no-binary pydantic
+RUN rm -rf /tmp/raster
+
+# Remove system dependencies
+RUN yum remove -y gcc-c++
+
+# Reduce package size and remove useless files
+RUN cd /asset && find . -type f -name '*.pyc' | while read f; do n=$(echo $f | sed 's/__pycache__\///' | sed 's/.cpython-[0-9]*//'); cp $f $n; done;
+RUN cd /asset && find . -type d -a -name '__pycache__' -print0 | xargs -0 rm -rf
+RUN cd /asset && find . -type f -a -name '*.py' -print0 | xargs -0 rm -f
+RUN find /asset -type d -a -name 'tests' -print0 | xargs -0 rm -rf
+RUN rm -rdf /asset/numpy/doc/ /asset/boto3* /asset/botocore* /asset/bin /asset/geos_license /asset/Misc
+
+COPY handlers/raster_handler.py /asset/handler.py
+
+CMD ["echo", "hello world"]
