+++ cdk/runtimes/eoapi/raster/README.md
@@ -1,0 +1,3 @@
+## eoapi.raster
+
+![](https://user-images.githubusercontent.com/10407788/151455911-c455a043-3313-4c26-b980-042cb80787a3.png)
