+++ cdk/runtimes/eoapi/raster/eoapi/raster/main.py
@@ -2,29 +2,8 @@
 Handler for AWS Lambda.
 """
 
-import asyncio
-import os
-
-from utils import get_secret_dict
-
-# Update postgres env variables before importing titiler.pgstac.main
-pgstac_secret_arn = os.environ["PGSTAC_SECRET_ARN"]
-
-secret = get_secret_dict(pgstac_secret_arn)
-os.environ.update(
-    {
-        "postgres_host": secret["host"],
-        "postgres_dbname": secret["dbname"],
-        "postgres_user": secret["username"],
-        "postgres_pass": secret["password"],
-        "postgres_port": str(secret["port"]),
-    }
-)
-
-
 from fastapi import APIRouter, Request
 from fastapi.responses import RedirectResponse
-from mangum import Mangum
 from rio_tiler.io import STACReader
 from titiler.core.factory import MultiBaseTilerFactory, TilerFactory
 from titiler.extensions import (
@@ -32,16 +11,9 @@ from titiler.extensions import (
     cogViewerExtension,
     stacViewerExtension,
 )
-from titiler.pgstac.db import connect_to_db  # noqa: E402
 from titiler.pgstac.main import app  # noqa: E402
 
-from factory import MosaicTilerFactory
-
-
-@app.on_event("startup")
-async def startup_event() -> None:
-    """Connect to database on startup."""
-    await connect_to_db(app)
+from eoapi.raster.factory import MosaicTilerFactory
 
 
 ########################################
@@ -103,11 +75,3 @@ async def redirect_to_searches(request: Request):
 
 
 app.include_router(redirect_router)
-
-
-handler = Mangum(app, lifespan="off")
-
-
-if "AWS_EXECUTION_ENV" in os.environ:
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(app.router.startup())
