+++ cdk/runtimes/eoapi/raster/eoapi/raster/models.py
