+++ cdk/runtimes/eoapi/raster/pyproject.toml
@@ -1,0 +1,61 @@
+[project]
+name = "eoapi.raster"
+description = "Custom raster tiling service for MAAP"
+readme = "README.md"
+requires-python = ">=3.8"
+authors = [
+    {name = "Vincent Sarago", email = "vincent@developmentseed.com"},
+    {name = "Henry Rodman", email = "henry@developmentseed.com"},
+]
+license = {text = "MIT"}
+classifiers = [
+    "Intended Audience :: Information Technology",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Topic :: Scientific/Engineering :: GIS",
+]
+dynamic = ["version"]
+dependencies = [
+    "titiler.pgstac==1.3.1",
+    "titiler.extensions[cogeo]==0.18.5",
+    "titiler.mosaic==0.18.5",
+    "titiler.pgstac==1.3.1",
+    "pystac-client==0.8.4",
+    "stac-pydantic==3.1.3",
+    "importlib_resources>=1.1.0;python_version<'3.9'",
+]
+
+[project.optional-dependencies]
+# https://www.psycopg.org/psycopg3/docs/api/pq.html#pq-module-implementations
+psycopg = [  # pure python implementation
+    "psycopg[pool]"
+]
+psycopg-c = [  # C implementation of the libpq wrapper
+    "psycopg[c,pool]"
+]
+
+psycopg-binary = [  # pre-compiled C implementation
+    "psycopg[binary,pool]"
+]
+test = [
+    "pytest",
+    "pytest-cov",
+    "pytest-asyncio",
+    "httpx",
+]
+
+[build-system]
+requires = ["pdm-pep517"]
+build-backend = "pdm.pep517.api"
+
+[tool.pdm.version]
+source = "file"
+path = "eoapi/raster/__init__.py"
+
+[tool.pdm.build]
+includes = ["eoapi/raster"]
+excludes = ["tests/", "**/.mypy_cache", "**/.DS_Store"]
