--- cdk/runtimes/titiler-pgstac-api/dev_requirements.txt
@@ -1,2 +1,0 @@
-uvicorn
-boto3>=1.26.139
