--- cdk/runtimes/titiler-pgstac-api/requirements.txt
@@ -1,6 +1,0 @@
-titiler.core==0.18.5
-titiler.extensions[cogeo]==0.18.5
-titiler.mosaic==0.18.5
-titiler.pgstac==1.3.1
-psycopg[binary, pool]
-pystac-client==0.8.4
